+++ test/chisel_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:chisel/chisel.dart';
+import 'package:chisel/src/annotations.dart';
 import 'package:postgres/postgres.dart';
 import 'package:test/test.dart';